moved get_cql_dispatcher tests to test_models_cass.py
