+++ otter/test/test_effect_dispatcher.py
@@ -2,19 +2,13 @@
 
 from effect import Constant, Delay, Effect, sync_perform
 from effect.ref import ReadReference, Reference
-from effect.twisted import deferred_performer
 
-import mock
-
-from twisted.internet.defer import succeed
 from twisted.trial.unittest import SynchronousTestCase
 
 from otter.auth import Authenticate, InvalidateToken
 from otter.cloud_client import TenantScope
 from otter.effect_dispatcher import (
-    get_cql_dispatcher, get_full_dispatcher, get_legacy_dispatcher,
-    get_simple_dispatcher)
-from otter.models.cass import CQLQueryExecute
+    get_full_dispatcher, get_legacy_dispatcher, get_simple_dispatcher)
 from otter.models.intents import GetScalingGroupInfo
 from otter.util.pure_http import Request
 from otter.util.retry import Retry
@@ -97,28 +91,3 @@ class FullDispatcherTests(SynchronousTestCase, IntentSupportMixin):
 
     def get_intents(self):
         return full_intents()
-
-
-class CQLDispatcherTests(SynchronousTestCase):
-    """Tests for :func:`get_cql_dispatcher`."""
-
-    def test_intent_support(self):
-        """Basic intents are supported by the dispatcher."""
-        dispatcher = get_simple_dispatcher(None)
-        for intent in simple_intents():
-            self.assertIsNot(dispatcher(intent), None)
-
-    @mock.patch('otter.effect_dispatcher.perform_cql_query')
-    def test_cql_disp(self, mock_pcq):
-        """The :obj:`CQLQueryExecute` performer is called."""
-
-        @deferred_performer
-        def performer(c, d, i):
-            return succeed('p' + c)
-
-        mock_pcq.side_effect = performer
-
-        dispatcher = get_cql_dispatcher(object(), 'conn')
-        intent = CQLQueryExecute(query='q', params='p', consistency_level=1)
-        eff = Effect(intent)
-        self.assertEqual(sync_perform(dispatcher, eff), 'pconn')
